Stage 2 AI
Fixed some spacing on comments

Set the playing board print statements to be more readable instead of
squishing them

Added first stage of AI. This stage checks board for hits. If a hit is
found, it checks the surrounding spaces for hits as well. If one of
those spaces has not been attempted, it attempts it. If all spaces
around the hit have been tried, it tried a random spot. Repeat for each
move.
The limitation here is that it checks ALL spaces surrounding the hit,
which human players normally wouldn't do. Next stage will be to
recognize rows of hits (so, check hit for surrounding hits first, then
move to trying the other places).

There is a LOT of repeated code in the AI code right now. However, I
wanted to get this pushed because it works. Going to refactor next
before moving on to the next stage of the AI.

Also added a test mode for the AI to play against itself because playing
against it real time takes too long. This is called "watch" mode.

Also found a bug in setting up the player board (says positions are
invalid when they should be valid) but that's going to be tackled during
the refactoring.

Fixed bug where if player chose move they'd already tried, AI gets extra
move (as player's move was invalid)

Seriously so much repeated code but it's late.